tensorflow_object_detection_ssd_coco/requirements.txt
absl-py==0.11.0
appdirs==1.4.3
astunparse==1.6.3
backcall==0.2.0
CacheControl==0.12.6
cachetools==4.2.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lockfile==0.12.2
Markdown==3.3.3
matplotlib==3.3.3
msgpack==0.6.2
numpy==1.19.4
oauthlib==3.1.0
object-detection==0.1
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.3
parso==0.8.1
pathlib==1.0.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
progress==1.5
prompt-toolkit==3.0.9
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
tf-slim==1.1.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
